增加事物

+++ project-contribute-server/src/main/java/com/camelot/pmt/service/impl/SysUserServiceImpl.java
@@ -2,8 +2,10 @@ package com.camelot.pmt.service.impl;
 
 import java.util.Date;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.camelot.pmt.mapper.SysUserMapper;
 import com.camelot.pmt.model.SysUser;
@@ -26,7 +28,14 @@ public class SysUserServiceImpl implements SysUserService {
      */
     @Override
     public SysUser selectByPrimaryKey(Integer id) {
-        return sysUserMapper.selectByPrimaryKey(id);
+        if (id == null) {
+            return null;
+        }
+        SysUser sysUser = sysUserMapper.selectByPrimaryKey(id);
+        if (sysUser != null) {
+            return sysUser;
+        }
+        return null;
     }
 
     /**
@@ -36,7 +45,11 @@ public class SysUserServiceImpl implements SysUserService {
      * @return
      */
     @Override
+    @Transactional
     public int deleteByPrimaryKey(Integer id) {
+        if (id == null) {
+            return 0;
+        }
         return sysUserMapper.deleteByPrimaryKey(id);
     }
 
@@ -53,8 +66,12 @@ public class SysUserServiceImpl implements SysUserService {
      * @return
      */
     @Override
+    @Transactional
     public int insert(String userName, String realName, String password, String email, String tel, String userDesc,
-            Integer state) {
+                      Integer state) {
+        if (StringUtils.isEmpty(userName) || StringUtils.isEmpty(password) || state == null) {
+            return 0;
+        }
         // 创建人修改人应从session中获取暂留
         return sysUserMapper.insert(userName, realName, password, email, tel, userDesc, state, 1, new Date(), 1,
                 new Date());
@@ -67,6 +84,7 @@ public class SysUserServiceImpl implements SysUserService {
      * @return
      */
     @Override
+    @Transactional
     public int updateByPrimaryKeySelective(SysUser sysUser) {
         return sysUserMapper.updateByPrimaryKeySelective(sysUser);
     }
